Moved some widget components to the widget folder
