+++ lib/screens/logIn.dart
@@ -4,12 +4,12 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:verbatim_frontend/Components/my_textfield.dart';
+import 'package:verbatim_frontend/widgets/my_textfield.dart';
 import 'package:verbatim_frontend/screens/forgotPassword.dart';
 import 'package:verbatim_frontend/screens/signUp.dart';
 import 'package:verbatim_frontend/screens/signupErrorMessage.dart';
-import '../Components/my_button.dart';
-import '../Components/my_button_no_image.dart';
+import '../widgets/my_button.dart';
+import '../widgets/my_button_no_image.dart';
 import 'draft.dart';
 import 'globalChallenge.dart';
 import 'package:http/http.dart' as http;
