+++ lib/screens/signUp.dart
@@ -4,11 +4,11 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:verbatim_frontend/Components/my_button_no_image.dart';
-import 'package:verbatim_frontend/Components/my_textfield.dart';
+import 'package:verbatim_frontend/widgets/my_button_no_image.dart';
+import 'package:verbatim_frontend/widgets/my_textfield.dart';
 import 'package:verbatim_frontend/screens/logIn.dart';
 import 'package:verbatim_frontend/screens/signupErrorMessage.dart';
-import '../Components/my_button.dart';
+import '../widgets/my_button.dart';
 import 'globalChallenge.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
